requirements.txt
fastapi[all]
python-dotenv
scikit-learn